Extract request methods into request mixin
